+++ options/db.lua
@@ -188,6 +188,10 @@ function SAO.LoadDB(self)
         end
     end
 
+    if not db.questions then
+        db.questions = {};
+    end
+
     -- Migration from older versions
     if not db.version or db.version < 091 then
         migrateTo091(db);
@@ -218,11 +222,64 @@ function SAO.LoadDB(self)
     end
 end
 
+-- Ask questions after the database is loaded
+function SAO.AskQuestions(self)
+    local displayGameSaoVar = "displaySpellActivationOverlays";
+    if SAO.IsProject(SAO.MOP_AND_ONWARD) -- Issue starts with Mists of Pandaria
+    and C_CVar.GetCVarInfo(displayGameSaoVar) ~= nil -- Can only operate if the variable is supported
+    and C_CVar.GetCVarBool(displayGameSaoVar) -- Bother asking only if the game alert is enabled
+    and SpellActivationOverlayDB.alert.enabled -- Only ask if the alert is enabled
+    and (
+        SpellActivationOverlayDB.questions.disableGameAlert == nil -- Ask if the user has not answered yet
+        or SpellActivationOverlayDB.questions.disableGameAlert == "yes" -- Ask again if the user already answered "yes"
+        )
+    then
+        local optionSequence = string.format(
+            "%s > %s > %s",
+            OPTIONS,
+            COMBAT_LABEL,
+            SPELL_ALERT_OPACITY
+        );
+        StaticPopupDialogs["SAO_DISABLE_GAME_ALERT"] = {
+            text = "",
+            button1 = YES,
+            button2 = NO,
+            OnShow = function(self)
+                if self.data.answered == "yes" then
+                    -- Player already answered "yes" but the option came back
+                    -- This can happen if the player disabled the game's spell alert, then re-enabled it
+                    self.text:SetText(SAO:spellAlertConflictsAgain());
+                else
+                    -- Player has not answered yet
+                    self.text:SetText(SAO:spellAlertConflicts());
+                end
+            end,
+            OnAccept = function(self)
+                SetCVar(displayGameSaoVar, false);
+                SpellActivationOverlayDB.questions.disableGameAlert = "yes";
+                SAO:Info(Module, SAO:gameSpellAlertsDisabled().."\n"..SAO:gameSpellAlertsChangeLater(optionSequence));
+            end,
+            OnCancel = function(self)
+                SpellActivationOverlayDB.questions.disableGameAlert = "no";
+                SAO:Info(Module, SAO:gameSpellAlertsLeftAsIs().."\n"..SAO:gameSpellAlertsChangeLater(optionSequence));
+            end,
+            whileDead = true,
+            customAlertIcon = "Interface/Addons/SpellActivationOverlay/textures/rkm128",
+            hideOnEscape = true,
+            noCancelOnEscape = true,
+            timeout = 0,
+            preferredindex = STATICPOPUP_NUMDIALOGS
+        };
+        StaticPopup_Show("SAO_DISABLE_GAME_ALERT", nil, nil, { answered = SpellActivationOverlayDB.questions.disableGameAlert });
+    end
+end
+
 -- Utility frame dedicated to react to variable loading
 local loader = CreateFrame("Frame", "SpellActivationOverlayDBLoader");
 loader:RegisterEvent("VARIABLES_LOADED");
 loader:SetScript("OnEvent", function (self, event)
     SAO:LoadDB();
+    SAO:AskQuestions();
     SAO:ApplyAllVariables();
     SpellActivationOverlayOptionsPanel_Init(SAO.OptionsPanel);
     loader:UnregisterEvent("VARIABLES_LOADED");